requirements.txt
langchain==0.0.227
openai==0.27.8
argparse==1.4.0
python-dotenv==1.0.0
urllib3==1.26.6
unidiff==0.7.5
pydantic==1.10.11
requests==2.31.0
chromadb==0.3.29
sentence-transformers==2.2.2